Internalized some of the type extensions only meant to be used by FA
