+++ Tests/FluentAssertions.Specs/Exceptions/ExceptionAssertionSpecs.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
+using FluentAssertions.Specs.Common;
 using Xunit;
 using Xunit.Sdk;
