+++ Tests/FluentAssertions.Specs/Specialized/TaskOfTAssertionSpecs.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using FluentAssertions.Extensions;
+using FluentAssertions.Specs.Common;
 using Xunit;
 using Xunit.Sdk;
